Moved original first unit test to test_statement.
