+++ tests/unit/test_statement.py
@@ -9,6 +9,16 @@ def assert_beautified(result, expected):
     assert jsbeautifier.beautify(result) == jsbeautifier.beautify(expected)
 
 
+def test_binop_statement():
+    binop = ast.BinOp(
+        left=ast.Constant(value=1),
+        op=ast.Add(),
+        right=ast.Constant(value=2)
+    )
+    result = Main('').process_statement(binop)
+    assert result == "1 + 2"
+
+
 def test_assign_statement():
     assign = ast.Assign(
         targets=[ast.Name(id='x')],
